% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_cla_ll_threaded}
\alias{calc_cla_ll_threaded}
\title{cla log likelihood using tbb flow}
\usage{
calc_cla_ll_threaded(
  ances,
  states_R,
  forTime_R,
  lambdas_R,
  mus_R,
  Q,
  num_threads = 1L,
  method = "odeint::bulirsch_stoer",
  is_complete_tree = FALSE
)
}
\arguments{
\item{ances}{ances}

\item{states_R}{states}

\item{Q}{q}

\item{num_threads}{num threads}

\item{method}{integration method}

\item{is_complete_tree}{is complete tree?}

\item{forTime_r}{ff}

\item{lambdas}{list of lambdas}

\item{mus}{vector of mus}
}
\description{
cla loglik
}
