% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_ll_threaded}
\alias{calc_ll_threaded}
\title{ll threaded}
\usage{
calc_ll_threaded(ll, mm, Q, ances, for_time, states, num_threads)
}
\arguments{
\item{ll}{lambdas}

\item{mm}{mus}

\item{Q}{qs}

\item{ances}{vector of ances}

\item{for_time}{fortime}

\item{states}{states matrix}

\item{node_M}{out}
}
\value{
log likelihood
}
\description{
ll threaded
}
