% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cla_calThruNodes_cpp}
\alias{cla_calThruNodes_cpp}
\title{function to calculate log likelihood using cpp}
\usage{
cla_calThruNodes_cpp(ances, states_R, forTime_R, lambdas, mus, Q)
}
\arguments{
\item{ances}{vector of ancestors}

\item{states_R}{numericmatrix of states}

\item{forTime_R}{numericmatrix of phylogeny information}

\item{lambdas}{vector of lambdas}

\item{mus}{vector of mus}

\item{Q}{numericmatrix with transition rates}
}
\description{
function to calculate log likelihood using cpp
}
