% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cla_calThruNodes_store_cpp}
\alias{cla_calThruNodes_store_cpp}
\title{function to do cpp stuff}
\usage{
cla_calThruNodes_store_cpp(
  ances,
  states_R,
  forTime_R,
  lambdas,
  mus,
  Q,
  dt,
  is_complete_tree
)
}
\arguments{
\item{ances}{ances}

\item{states_R}{states_R}

\item{forTime_R}{fr}

\item{lambdas}{l}

\item{mus}{mus}

\item{Q}{Q}

\item{dt}{dt}

\item{is_complete_tree}{ss}
}
\description{
function to do cpp stuff
}
