% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cla_secsse_sim.R
\name{cla_secsse_sim}
\alias{cla_secsse_sim}
\title{Function to simulate a tree, conditional on observing all states.}
\usage{
cla_secsse_sim(
  states,
  lambdas,
  mus,
  timeSimul,
  qs,
  pool_init_states,
  maxSpec,
  the_traits_to_sim
)
}
\arguments{
\item{states}{states}

\item{lambdas}{speciation rates}

\item{mus}{extinction rates}

\item{timeSimul}{Crown age of the tree}

\item{qs}{The Q matrix, for example the result of function q_doubletrans.}

\item{pool_init_states}{I don't know what this is?}

\item{maxSpec}{Maximum number of species in the tree (please note that the
tree is not conditioned on this number, but that this is a safeguard against
generating extremely large trees).}

\item{the_traits_to_sim}{the traits to simulate}
}
\value{
a list with three properties: phy: Phylogenies,
traits: the traits and thirdly: initialState, delineating the
initial state.
}
\description{
Function to simulate a tree, conditional on observing all states.
}
