% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secsse_lambdas.R
\name{create_default_q_list}
\alias{create_default_q_list}
\title{helper function to create a default q_matrix list}
\usage{
create_default_q_list(
  state_names = c("0", "1"),
  num_concealed_states,
  lambda_matrices = NULL
)
}
\arguments{
\item{state_names}{names of the observed states}
}
\description{
This function generates a generic transition list
}
