% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secsse_prep.R
\name{create_default_shift_matrix}
\alias{create_default_shift_matrix}
\title{helper function to create a default shift_matrix list}
\usage{
create_default_shift_matrix(
  state_names = c("0", "1"),
  num_concealed_states = 2,
  mus = NULL
)
}
\arguments{
\item{state_names}{names of the observed states}

\item{num_concealed_states}{number of concealed states}

\item{mus}{previously defined mus - used to choose indicator number}
}
\description{
This function generates a generic shift matrix to be used with the function
create_transition_matrix.
}
\examples{
shift_matrix <- create_default_shift_matrix(state_names = c(0, 1),
                                            num_concealed_states = 2,
                                            mus = c(1, 2, 1, 2))
q_matrix <- create_transition_matrix(state_names = c(0, 1),
                                     num_concealed_states = 2,
                                     shift_matrix = shift_matrix,
                                     diff.conceal = FALSE)
}
