% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secsse_prep.R
\name{create_default_transition_list}
\alias{create_default_transition_list}
\title{helper function to create a default transition list}
\usage{
create_default_transition_list(
  state_names = c("0", "1"),
  consider_combinations = FALSE
)
}
\arguments{
\item{state_names}{names of the observed states}

\item{consider_combinations}{should there be extra
states considering combinations of the observed states?}
}
\description{
This function generates a generic transition list, assuming no transitions
between states, e.g. a species of observed state 0 generates two daughter
species with state 0 as well.
If consider_combinations is set to TRUE, the function automatically
generates transitions for the combined states as well, for example if there
are two states A and B, it adds a third state AB, which generates two unique
daughter species, in this case each having either trait A, or B.
}
