% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secsse_prep.R
\name{create_mu_vector}
\alias{create_mu_vector}
\title{function to generate mus vector}
\usage{
create_mu_vector(state_names, num_concealed_states, model = "CR", lambda_list)
}
\arguments{
\item{state_names}{names of the observed states}

\item{num_concealed_states}{number of concealed states}

\item{model}{model replicated, available are "CR", "ETD" and "CTD"}

\item{lambda_list}{previously generated list of lambda matrices, 
used to infer the rate number to start with}
}
\value{
mu vector
}
\description{
function to generate mus vector
}
