% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secsse_lambdas.R
\name{create_transition_matrix}
\alias{create_transition_matrix}
\title{helper function to neatly setup a Q matrix, without transitions to 
concealed states (only observed transitions shown)}
\usage{
create_transition_matrix(state_names, transition_list)
}
\arguments{
\item{state_names}{names of observed states}

\item{transition_list}{matrix of transitions, indicating in order: 1) 
starting state (typically the column in the transition matrix), 2) ending 
state (typically the row in the transition matrix) and 3) associated rate 
indicator}
}
\value{
transition matrix
}
\description{
helper function to neatly setup a Q matrix, without transitions to 
concealed states (only observed transitions shown)
}
