% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secsse_prep.R
\name{expand_q_matrix}
\alias{expand_q_matrix}
\title{function to expand an existing q_matrix to a number of
concealed states, highly similar to q_doubletrans.}
\usage{
expand_q_matrix(q_matrix, num_concealed_states, diff.conceal = FALSE)
}
\arguments{
\item{q_matrix}{q_matrix with only transitions between observed states}

\item{num_concealed_states}{number of concealed states}

\item{diff.conceal}{should we use the same number of rates for the
concealed state transitions, or should all concealed state transitions
have separate rates? Typically, FALSE is fine and should be used in order
to avoid having a huge number of parameters.}
}
\value{
updated q matrix
}
\description{
function to expand an existing q_matrix to a number of
concealed states, highly similar to q_doubletrans.
}
