% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secsse_lambdas.R
\name{fill_in}
\alias{fill_in}
\title{helper function to enter parameter value on their right place}
\usage{
fill_in(object, params)
}
\arguments{
\item{object}{lambda matrices, q_matrix or mu vector}

\item{params}{parameters in order, where each value reflects the value
of the parameter at that position, e.g. c(0.3, 0.2, 0.1) will fill out
the value 0.3 for the parameter with rate indentifier 1, 0.2 for the
parameter with rate identifier 2 and 0.1 for the parameter with rate
identifier 3}
}
\description{
helper function to enter parameter value on their right place
}
