% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_state_AB.R
\name{plot_state_AB}
\alias{plot_state_AB}
\title{function to plot the hidden-states across a tree}
\usage{
plot_state_AB(focal_tree, ancestral_states)
}
\arguments{
\item{focal_tree}{the used phylogeny}

\item{ancestral_states}{ancestral states matrix, as returned by 
cla_secsse_loglik}
}
\value{
ggplot object
}
\description{
This plots the probability of observing hidden state A along the
tree. Please note that this takes the probabilities at the nodes and 
interpolates the probabilities along the branches. Furthermore, it assumes
the tips have 50/50 probability of either state.
}
