% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_state_exact_old.R
\name{plot_state_exact_old}
\alias{plot_state_exact_old}
\title{function to extract local likelihoods along all branches}
\usage{
plot_state_exact_old(
  parameters,
  focal_tree,
  traits,
  num_concealed_states,
  states,
  dt,
  res
)
}
\arguments{
\item{parameters}{used parameters for the likelihood caluclation}

\item{focal_tree}{used phylogeny}

\item{traits}{used traits}

\item{num_concealed_states}{number of concealed states}

\item{states}{the states matrix returned by cla_secsse_loglik}

\item{dt}{chosen time step}

\item{res}{resolution of the color scale used.}
}
\value{
nothing
}
\description{
this function will evaluate the log likelihood locally along
all branches and plot the result, using the phytools package.
}
