% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_state_AB.R
\name{plot_state_traits}
\alias{plot_state_traits}
\title{function to plot the observed states across a tree}
\usage{
plot_state_traits(focal_tree, traits, ancestral_states)
}
\arguments{
\item{focal_tree}{the used phylogeny}

\item{ancestral_states}{ancestral states matrix, as returned by 
cla_secsse_loglik}
}
\value{
list with as many plots as traits
}
\description{
This plots the probability of observing each tip state across
the tree. Coloring is interpolated between known probabilities at the nodes.
Shown are separate plots for each trait.
}
