% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secsse_utils.R
\name{prepare_full_lambdas}
\alias{prepare_full_lambdas}
\title{Prepares the entire set of lambda matrices for cla_secsse.}
\usage{
prepare_full_lambdas(traits, num_concealed_states, lambd_and_modeSpe)
}
\arguments{
\item{traits}{vector with trait states, order of states must be the same as tree tips, for help, see vignette.}

\item{num_concealed_states}{number of concealed states, generally equivalent to number of examined states.}

\item{lambd_and_modeSpe}{a matrix with the 4 models of speciation possible.}
}
\value{
A list of lambdas, its length would be the same as the number of trait states * num_concealed_states..
}
\description{
It provides the set of matrices containing all the speciation rates
}
