% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secsse-package.R
\docType{package}
\name{secsse-package}
\alias{secsse}
\alias{secsse-package}
\title{secsse: Several Examined and Concealed States-Dependent Speciation and Extinction}
\description{
Simultaneously infers state-dependent diversification across two or more states of a single or multiple traits while accounting for the role of a possible concealed trait. See Herrera-Alsina et al. (2019) \doi{10.1093/sysbio/syy057}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://rsetienne.github.io/secsse/}
  \item \url{https://github.com/rsetienne/secsse}
  \item Report bugs at \url{https://github.com/rsetienne/secsse/issues}
}

}
\author{
\strong{Maintainer}: Rampal S. Etienne \email{r.s.etienne@rug.nl} (\href{https://orcid.org/0000-0003-2142-7612}{ORCID})

Authors:
\itemize{
  \item Rampal S. Etienne \email{r.s.etienne@rug.nl} (\href{https://orcid.org/0000-0003-2142-7612}{ORCID})
  \item Leonel Herrera Alsina \email{leonelhalsina@gmail.com} (\href{https://orcid.org/0000-0003-0474-3592}{ORCID})
  \item Paul van Els \email{paulvanels@gmail.com} (\href{https://orcid.org/0000-0002-9499-8873}{ORCID})
}

Other contributors:
\itemize{
  \item Thijs Janzen \email{t.janzen@rug.nl} (\href{https://orcid.org/0000-0002-4162-1140}{ORCID}) [contributor]
  \item Hanno Hildenbrandt \email{h.hildenbrandt@rug.nl} (\href{https://orcid.org/0000-0002-6784-1037}{ORCID}) [contributor]
  \item Pedro Santos Neves \email{p.m.santos.neves@rug.nl} (\href{https://orcid.org/0000-0003-2561-4677}{ORCID}) [contributor]
}

}
\keyword{internal}
