% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secsse_sim_cond_cpp.R
\name{secsse_sim_cond_cpp}
\alias{secsse_sim_cond_cpp}
\title{Function to simulate a tree, conditional on observing all states.}
\usage{
secsse_sim_cond_cpp(
  states,
  lambdas,
  mus,
  timeSimul,
  qs,
  pool_init_states = states,
  maxSpec
)
}
\arguments{
\item{states}{states}

\item{lambdas}{speciation rates}

\item{mus}{extinction rates}

\item{timeSimul}{Crown age of the tree}

\item{qs}{The Q matrix, for example the result of function q_doubletrans.}

\item{pool_init_states}{pool of initial states, in case this is different
from all available states.}

\item{maxSpec}{Maximum number of species in the tree (please note that the
tree is not conditioned on this number, but that this is a safeguard against
generating extremely large trees).}
}
\value{
a list with three properties: phy: Phylogenies,
traits: the traits and thirdly: initialState, delineating the
initial state.
}
\description{
Function to simulate a tree, conditional on observing all states.
}
