% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secsse_utils.R
\name{sortingtraits}
\alias{sortingtraits}
\title{Data checking and trait sorting
In preparation for likelihood calculation, it orders trait data according
the tree tips}
\usage{
sortingtraits(trait_info, phy)
}
\arguments{
\item{trait_info}{data frame where first column has species ids and the second
one is the trait associated information.}

\item{phy}{phylogenetic tree of class \code{phylo}, rooted and with
branch lengths. Alternatively, multiple phylogenetic trees can be provided
as the \code{multiPhylo} class.}
}
\value{
Vector of traits
}
\description{
Data checking and trait sorting
In preparation for likelihood calculation, it orders trait data according
the tree tips
}
\examples{
# Some data we have prepared
data(traits)
data('phylo_vignette')
traits <- sortingtraits(traits, phylo_vignette)
}
