% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secsse_data.R
\docType{data}
\name{timing_data}
\alias{timing_data}
\title{Computation time information}
\format{
A tibble
}
\usage{
timing_data
}
\description{
Results of benchmarking different versions of secsse
}
\keyword{datasets}
