% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{traitinfo}
\alias{traitinfo}
\title{A table with trait info to run the vignette}
\format{
A data frame where each species has a trait state associated
}
\description{
An example of trait information in the right format for secsse
}
