% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secsse_data.R
\docType{data}
\name{traits}
\alias{traits}
\title{A table with trait info to run the vignette}
\format{
A data frame where each species has a trait state associated
}
\usage{
traits
}
\description{
An example of trait information in the right format for secsse
}
\keyword{datasets}
