% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cla_calThruNodes_cpp}
\alias{cla_calThruNodes_cpp}
\title{function to do cpp stuff}
\usage{
cla_calThruNodes_cpp(
  ances,
  states_R,
  forTime_R,
  lambdas,
  mus,
  Q,
  method,
  atol,
  rtol,
  is_complete_tree
)
}
\arguments{
\item{ances}{ances}

\item{states_R}{states_R}

\item{forTime_R}{fr}

\item{lambdas}{l}

\item{mus}{mus}

\item{Q}{Q}

\item{method}{method}

\item{atol}{atol}

\item{rtol}{rtol}

\item{is_complete_tree}{ss}
}
\value{
A named list with:
  * Numeric vector with states
  * Numeric vector of length 1 with loglik
  * Numeric vector merge branch
  * Numeric vector nodeM
}
\description{
function to do cpp stuff
}
