% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secsse_lambdas.R
\name{create_lambda_matrices}
\alias{create_lambda_matrices}
\title{helper function to automatically create lambda matrices, based on input}
\usage{
create_lambda_matrices(state_names, num_concealed_states, transition_list)
}
\arguments{
\item{state_names}{vector of names of all observed states}

\item{num_concealed_states}{number of hidden states}

\item{transition_list}{a matrix containing a description of all speciation
events, where the first column indicates the source state, the second and
third column indicate the two daughter states, and the fourth column gives 
the rate indicator used. E.g.: ["SA", "S", "A", 1] for a trait state "SA"
which upon speciation generates two daughter species with traits "S" and "A",
where the number 1 is used as indicator for optimization of the likelihood.}
}
\description{
helper function to automatically create lambda matrices, based on input
}
