% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secsse_prep.R
\name{create_mu_vector}
\alias{create_mu_vector}
\title{Generate mus vector}
\usage{
create_mu_vector(state_names, num_concealed_states, model = "CR", lambda_list)
}
\arguments{
\item{state_names}{vector of names of all observed states.}

\item{num_concealed_states}{number of concealed states, generally equivalent
to the number of examined states in the dataset.}

\item{model}{used model, choice of \code{"ETD"} (Examined Traits Diversification),
\code{"CTD"} (Concealed Traits Diversification) or \code{"CR"} (Constant Rate).}

\item{lambda_list}{previously generated list of lambda matrices,
used to infer the rate number to start with.}
}
\value{
mu vector
}
\description{
Generate mus vector
}
