% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secsse_prep.R
\name{create_mus}
\alias{create_mus}
\title{function to generate generic mus vector}
\usage{
create_mus(state_names, num_concealed_states, model = "CR", lambdas)
}
\arguments{
\item{state_names}{full state names, including concealed states, for example
c("0A", "1A", "0B", "1B")}

\item{num_concealed_states}{number of concealed states}

\item{model}{model replicated, available are "CR", "ETD" and "CTD"}

\item{lambdas}{previously generated lambda matrices, used to infer the rate
number to start with}
}
\value{
mu vector
}
\description{
function to generate generic mus vector
}
