% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ct_condition}
\alias{ct_condition}
\title{function to condition}
\usage{
ct_condition(d, y, t, ll, mm, Q, method, atol, rtol)
}
\arguments{
\item{d}{dd}

\item{y}{yd}

\item{t}{td}

\item{ll}{ll}

\item{mm}{mm}

\item{Q}{qd}

\item{method}{method}

\item{atol}{atol}

\item{rtol}{rtol}
}
\description{
function to condition
}
