% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secsse_data.R
\docType{data}
\name{example_phy_GeoSSE}
\alias{example_phy_GeoSSE}
\title{A phylogeny with traits at the tips}
\format{
A phylogeny as created by GeoSSE (diversitree)
}
\usage{
example_phy_GeoSSE
}
\description{
An example phylogeny for testing purposes
}
\keyword{datasets}
