% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secsse_lambdas.R
\name{expand_q_matrix}
\alias{expand_q_matrix}
\title{function to expand an existing q_matrix to a number of 
concealed states}
\usage{
expand_q_matrix(q_matrix, num_concealed_states)
}
\arguments{
\item{q_matrix}{q_matrix with only transitions between observed states}

\item{num_concealed_states}{number of concealed states}
}
\value{
updated q matrix
}
\description{
function to expand an existing q_matrix to a number of 
concealed states
}
