% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secsse_data.R
\docType{data}
\name{phylo_vignette}
\alias{phylo_vignette}
\title{A phylogenetic reconstuction to run the vignette}
\format{
Phylogenetic tree in phy format, rooted, including branch lengths
}
\usage{
phylo_vignette
}
\description{
An example phylogeny in the right format for secsse
}
\keyword{datasets}
