% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secsse_sim.R
\name{secsse_sim}
\alias{secsse_sim}
\title{Function to simulate a tree, conditional on observing all states.}
\usage{
secsse_sim(
  lambdas,
  mus,
  qs,
  crown_age,
  pool_init_states = NULL,
  maxSpec,
  conditioning = "none"
)
}
\arguments{
\item{lambdas}{speciation rates, in the form of a list of matrices}

\item{mus}{extinction rates, in the form of a vector}

\item{qs}{The Q matrix, for example the result of function q_doubletrans, but
generally in the form of a matrix.}

\item{crown_age}{crown age of the tree, tree will be simulated conditional
on non-extinction and this crown age.}

\item{pool_init_states}{pool of initial states, in case this is different
from all available states, otherwise leave at NULL}

\item{maxSpec}{Maximum number of species in the tree (please note that the
tree is not conditioned on this number, but that this is a safeguard against
generating extremely large trees).}

\item{conditioning}{can be 'obs_states', 'true_states' or 'none', the tree is
simulated until one is generated that contains all observed states, all 
true states (e.g. obs x hidden states), or is always returned.}
}
\value{
a list with three properties: phy: reconstructed phylogeny,
traits: the traits in order of tip label and thirdly: 
initialState, delineating the initial state at the root used.
}
\description{
Function to simulate a tree, conditional on observing all states.
}
