% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secsse_sim.R
\name{secsse_sim}
\alias{secsse_sim}
\title{function to simulate a tree under the SecSSE model.}
\usage{
secsse_sim(timeSimul, states, lambdas, mus, qs, speciesTraits, maxSpec)
}
\arguments{
\item{timeSimul}{Crown age of the tree}

\item{states}{states}

\item{lambdas}{speciation rates}

\item{mus}{extinction rates}

\item{qs}{transition rates in the form of a Q matrix (see the output of
q_doubletrans)}

\item{speciesTraits}{initial species' traits.}

\item{maxSpec}{Maximum number of species in the tree (please note that the
tree is not conditioned on this number, but that this is a safeguard against
generating extremely large trees).}
}
\value{
a list with four components: 1) phy: the resulting phylogeny,
2) ltable the associated Ltable, 3) speciesID a list of species
labels and 4) speciesTraits a list of species' traits.
}
\description{
function to simulate a tree under the SecSSE model.
}
